feature/cmake-vcpkg-toolchain: Refactor tests to reuse SparkIntegration test class and env loader
