+++ tests/dataframe_writer.cpp
@@ -9,32 +9,10 @@
 #include "config.h"
 #include "dataframe.h"
 #include "writer.h"
+#include "spark_integration.h"
 
 using ::testing::ElementsAre;
 
-class SparkIntegrationTest : public ::testing::Test
-{
-protected:
-    static SparkSession *spark;
-
-    static void SetUpTestSuite()
-    {
-        spark = &SparkSession::builder()
-                     .master("localhost")
-                     .appName("SparkConnectCppGTest")
-                     .getOrCreate();
-    }
-
-    static void TearDownTestSuite()
-    {
-        if (spark)
-        {
-            spark->stop();
-        }
-    }
-};
-
-SparkSession *SparkIntegrationTest::spark = nullptr;
 
 TEST_F(SparkIntegrationTest, ParquetWrite)
 {
