+++ tests/env_loader.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <string>
+#include <iostream>
+#include <sstream>
+#include <fstream>
+#include <cstdlib>
+
+void load_env(const std::string& path);
